Integration of SqlCensoredContentStore service
Summary: Use the service SqlCensoredContentStore to populate the hashmap of blacklisted blobs at startup. Every get and put call on a censoredblob instance will check the hash key against the existing keys in the hashmap. It will display a suggestive error in case the hash key is censored, otherwise, it will follow the normal behavior

Reviewed By: ikostia

Differential Revision: D15762598

fbshipit-source-id: 03a2d19f06934e8207f5294140c38a80b67a3c3d