revisionstore: impl HistoryStore for all MutableHistoryStore
Summary:
Similarly to MutableDataStore, all MutableHistoryStore can provide a read
access to them, so let's enforce this at compile time.

The PythonMutableHistoryPack has empty implementations as the type is only
going to be used for writing and is present only as a temporary solution until
Rust mutable history store are used exclusively.

Reviewed By: kulshrax

Differential Revision: D15717840

fbshipit-source-id: b2b3242aba548f72fbe218738dc77c49b24813f5