mononoke: do not reuse bad hosts if sync timed out
Summary:
We used to only invalidate the peer process when we explicitly fail the sync,
not when we timeout. Among other things, this means that if we talk to an
overloaded/bad server, we'll spend *all* of our retries timing out on it,
which will cause the queue to grow large. Let's invalidate the process
more aggressively.

Reviewed By: farnz

Differential Revision: D15877934

fbshipit-source-id: ed7e3d37e56392fed7fd9a58cb6cb7632bd9a694