mononoke: do not rely on auto-increment in the sync job
Summary:
This is the first step towards not skipping bundles due to race conditions
in our queue logic.

Below is our hypothesis of why this might happen and how we are trying to fix
it.

```
T   |   Mononoke Server 1   |   Mononoke Server 2   |   Sync job
---------------------------------------------------------------------
1   |   start transaction   |                       |
---------------------------------------------------------------------
2   |                       |   start transaction   |
---------------------------------------------------------------------
3   |   mysql decides auto  |                       |
    |   incremented key 20  |                       |
---------------------------------------------------------------------
4   |                       |   mysql decides auto  |
    |                       |   incremented key 21  |
---------------------------------------------------------------------
5   |                       |   INSERT              |
---------------------------------------------------------------------
6   |                       |   COMMIT              |
---------------------------------------------------------------------
7   |                       |                       |   read the
    |                       |                       |   latest key (21)
---------------------------------------------------------------------
8   |  INSERT               |                       |   update the latest
    |                       |                       |   key counter
---------------------------------------------------------------------
9   |  COMMIT               |                       |
---------------------------------------------------------------------

```

Note that in the above scenario key 20 will never be read by the sync job,
therefore bundle with this key will never be replayed.

To fix this and also to provide total order of all bookmark moves, we are
will not rely on the `AUTO_INCREMENT` mysql feature, but rather set the new
id ourelves, using the previously computed `MAX(id)` query. Note that it is
fine to insert a predetermined value into an `AUTO_INCREMENT` field.

Clearly, since `SELECT MAX(id)` and `INSERT INTO ...` are two different
queries, the behavior as introduced in this diff is racy, e.g. some other
bookmark move might've already used the `MAX(id)` key we've computed before
our insert succeeds. There are two aspects of this:
- how dangerous it is? In such case transaction commit fails and we fall back
  to pushrebase retry logic. We cannot introduce data inconsistency this way.
- what performance hit it introduces? We do not know, maybe it causes many more
  retries to happen. To address this potential hit, the following diff will
  make sure that when we have reasons to believe that the failure is because of
  the mentioned race, we only retry the database transaction, not the entire
  pushrebase.

Always winning this race (potentially after a few retries) is exactly what we want
to ensure that we have a total order of all the bookmark moves.

NB: I will not land this diff before introducing the next one

Reviewed By: krallin

Differential Revision: D15821895

fbshipit-source-id: cfb8180d1f43caa9282f06d7dd6a55fd856b71b8