mononoke: introduce a conditional retry function to the bookmarks crate
Summary:
See bottom diff for more context on the problem we are trying to solve.

This function can be used to use some more complex decision-making logic than
just `attempt < retry_num` for whether to retry or not. This also does not
have any delay, since in the envisioned use-case delays are not needed.

NB: `#[allow(dead_code)]` will be removed in the later diff. This is just for
the sake of diff separation.

NB 2: I think we should extract all of the different `retry` implementations into a separate crate, but I don't want to do it as part of this task.

Reviewed By: krallin

Differential Revision: D15855496

fbshipit-source-id: a47ccdb507634b202feecfa7f6076decd9451f18