commitcloud: don't record backed up heads for failed transactions
Summary:
When cloud sync pulls in new commits from commitcloud, it also adds them to the
backedup heads, so that the next backup doesn't try to push them to the server
again.  However, the pull happens in a transaction, which may fail.  If it does
fail, the backedup heads are still recorded.

Update the write to the backedup heads file to take place inside the
transaction in this case, so the update only happens if the transaction
succeeds.

If any unavailable heads have been recorded, ignore them.

Reviewed By: mitrandir77

Differential Revision: D15898056

fbshipit-source-id: e7affce1b10f507e2ee924d9bc78000c835ccd65