mononoke: fix filenode generation if one filenode is ancestor of another
Summary:
If a file has two parent filenodes and one of them is ancestor of another then
we want to keep only the descendant filenode as a parent.

(NOTE) Note that this diff doesn't fix all the corner cases. In integration tests a file that has
two parents is modified in a merge commit. Note that if it wasn't modified in
a merge commit then Mercurial produces a hash different from Mononoke. I'll
investigate why it happens

(NOTE) We had incorrect hashes in our test fixtures - this diff fixes them as well

Reviewed By: farnz

Differential Revision: D15896735

fbshipit-source-id: ea31071bc69fab02935887c665f6d03b64d5c572