edenapi: Add new ApiError type
Summary: This diff adds a new `ApiError` type that will serve as the general error type for Eden API operations. This error type is a based on the ["Error and ErrorKind pair" pattern](https://rust-lang-nursery.github.io/failure/error-errorkind.html) described in the `failure` crate's documentation. It uses a `failure::Context` to wrap an underlying error, and provides both the ability to provide users with high-quality custom error messages as well as the ability to programmatically consume such errors via pattern matching. See the doc comment in the code for usage details.

Reviewed By: xavierd

Differential Revision: D15913955

fbshipit-source-id: cb61f8c40dc04c78a5c281535163788e95c4f4e8