edenapi: add new errors module
Summary: Add a new `errors` module with a new `ApiResult` type. For now, this type is just an alias for `failure::Fallible`. The goal of this diff is to simply plumb the new result type throughout the crate. In a later diff, we will then leverage this to change the error type for the whole crate without needing to update all the function signatures.

Reviewed By: xavierd

Differential Revision: D15913956

fbshipit-source-id: be3c64051c67722b517add3c006006bcd9116962