windows: add a patch to make python.exe search for argv[0] path
Summary:
This makes it possible for `python setup.py` to import `distutils_rust` without
changing sys.path. The `argv[0]` path being part of `sys.path` is the expected
Python interpreter behavior.

Note: the `argv[0]` path is used in hgmain for POSIX, but not Windows because
of `0005-PC-aggressively-simplify-sys.path.patch`. That is desirable since on
Windows, the "main exe path" is the `argv[0]` path, and it already exists in
the default `sys.path`.

Reviewed By: ikostia

Differential Revision: D15935893

fbshipit-source-id: a759440ef3103dfdd49d97727ea74d09afe1d881