mononoke: make test-hash-validation.t deterministically pass
Summary:
The failures are due to Mercurial switching from Python 2.7.5 (system) to
2.7.14 (fbcode) as part of D15844921.

As per https://docs.python.org/2.7/library/stdtypes.html#dict.items, we can't
rely on dict traversal order and while preparing `getfiles` request on the
client side we do so many times.

Note also that Mercurial error does not really report the actual corrupted
file, rather the file that happens to be next in the requested queue when
server closes the connection, so version-dependent order breaks everything.

Reviewed By: krallin

Differential Revision: D15942912

fbshipit-source-id: c2ab064ee13cb7ab726a5d3a8459bf934b6c964b