types: speed up hex encode and decode
Summary:
Hex encoding and decoding is in the hot path for gettreepacks, particularly
when we have large fetches (for example with clones), we can easily do this in
the order of tens or hundreds of thousands.

Switch to using a SSE/AVX2 optimized hex encoding/decoding which results in
a 10x performance improvement for decoding. Encoding was already using a
relatively optimized version, so gain is expected to be much lower for encoding.

Reviewed By: farnz

Differential Revision: D15940209

fbshipit-source-id: 28734f45f7508a94b110e25c01e1baa955ebd4e4