fix the Windows build
Summary:
D15901794 deleted the move `ThriftLogHelper` move constructor, to suppress
some warnings from clang on Mac OS.  However, this causes build failures on
Windows, which needs the move constructor.  Clang does not need the move
constructor due to copy elision, which is guaranteed by C++17.

This also changes the itcLogger_ member variable to be a copy instead of a
reference: `folly::Logger` is just a simple wrapper around a
`folly::LogCategory*`, so making a copy should be just as efficient than using
a reference (if not more, since we no longer have to traverse the reference).

Reviewed By: pkaush

Differential Revision: D15936517

fbshipit-source-id: 92ab47ac9985a04e360f529bc99705a7cad5c5b2