simplecache: fix infinite loop when reading from memcache socket
Summary:
In S179901, we've seen a couple of instances of `hg status` consuming as much
memory as they could until getting OOM killed. Looking at backtraces for these
processes (P67246356) the code spins trying to read data from memcache. The
connection is however closed and therefore `recv` returns an empty string,
which is then added to a list and the `recv` is retried again.

Reviewed By: kulshrax

Differential Revision: D15947154

fbshipit-source-id: f1216e26b0ce159377ff8e82bedc13a7cc3d0444