mononoke: add lazy fetching of HgBlobEntry content
Summary:
Quite surprisingly, this diff saves quite a lot of memory on large gettreepack
requests - >100 Mb per request! So in total it might save gigabytes if a few
big gettreepack requests were sent at once.

This result is surprising and I found it almost accidentally. It's hard to
clearly explain why it helps with memory usage, but my current thinking is that
previously it eagerly fetched a lot of manifests and kept all of them in
memory.

That's not a great explanation, but the diff should be fairly safe to land
anyway.

Reviewed By: aslpavel, farnz

Differential Revision: D15959914

fbshipit-source-id: aee534a981dd79a52111a5123a71f90a6d5af79b