remotefilelog: add getpackv2 to the wireprotocol
Summary:
The getpackv1 API doesn't support LFS blobs as no metadata flags are sent over
the wire. This means that repositories like ovrsource can't use it, which
prevents these repos from moving away from loosefiles...

The getpackv2 is similar to getpackv1, except that it also sends the metadata
alongside the data. This means that the LFS flag is properly sent over the
wire.

Reviewed By: DurhamG

Differential Revision: D15905425

fbshipit-source-id: dab8dcb75c8d9db8c661f11f87318feca7d0f6a9