edenapi: Add conversion from ApiError to Python exceptions
Summary: Now that we have an error type that we can pattern match on, use this to raise an appropriate Python exception in the bindings. Note that due to Rust's coherence rules, we need to implement the conversion to Python errors in the `edenapi` crate. Since not all users will need this FFI functionality, it is gated behind a feature flag (which is enabled by default for Buck builds due to limitations in how Rust targets are specified).

Reviewed By: xavierd

Differential Revision: D15914529

fbshipit-source-id: 3f6a5a651fc62d2506231dd2a33fbc06528d2bce