blackbox: add pattern matching for filtering events
Summary:
Treat Event as a JSON value, then use flexible pattern matching to match it.
The matching logic is somewhat flexible (and powerful), but not designed to be
efficient, though. Practically, it's probably always a good idea to filter
events by a time range first, then filter by the (expensive) pattern matching.

Reviewed By: markbt

Differential Revision: D15685475

fbshipit-source-id: aae033c39dd6e1845818a6b738de2ed5b0139721