amend: do not rebase unrelated commits
Summary:
See the test change for what this change is about.

The new code has side effect on linearizing things for `<rebase_src>::`. This
is reflected on the second last test change. The new behavior is arguably more
desiable - The new code can fail if rebasing D onto C2 causes conflicts. If
that happens, and rebasing C+C2+D onto the new B does not cause conflicts, then
the old behavior is better. We can add fallback code to the old behavior if
there are conflicts later, by calling a plain `rebase` instead of `restack`.

Reviewed By: kulshrax

Differential Revision: D15978953

fbshipit-source-id: 20b9d62c6125ce2faf8e21bd86f6aad31ac38a0c