Moving Journal to Deque Structure
Summary:
Moving from the previous linked list approach to having a deque in the Journal.

The memory used by the deque is not tracked currently, the difference might be negligible though. I ran a script that touched a file 62 million times and here is the result comparing rss_memory used in eden and the journal's estimated memory: https://fburl.com/ods/6sq8g5vc and it still seems to closely estimate the right amount.

Reviewed By: strager

Differential Revision: D15944614

fbshipit-source-id: 6a1ac34ecd80c0eecb80411984f88f62ae712e91