remotefilelog: log exception type on ssh fallback
Summary: When Eden API falls back to SSH, log the exception type and message (in addition to the traceback) to the `hg_errors` table. Note that previously, we were attempting to log the error message but failing because we were using the wrong column name. (`msg` instead of `exception_msg`).

Reviewed By: xavierd

Differential Revision: D16014833

fbshipit-source-id: 06460574f66999b1293ea31b43b5c7ec89737144