edenapi: use http::StatusCode in ApiErrorKind::Http
Summary: Use the `StatusCode` type from the `http` crate to represent HTTP status codes rather than using a plain `u32`. This has the advantage of including the canonical description of the error in the `Display` output of the type, so the user doesn't need to look up the meaning of the particular HTTP status code.

Reviewed By: xavierd

Differential Revision: D16024749

fbshipit-source-id: f2325111d5ca647b193fe8c5fa4a270f890fdbee