stop hg svn rebuildmeta/updatemeta from talking to svn server
Summary:
# RFC, not sure if I know what I am doing but I know my goal as follow

In case of the entire `.hg/svn` missing, in order to get the subdir and uuid information,
hgsubversion needs to talk to svn server. In our cases, it's polluting our logging and
it's hard to know who is really using svn server.

In this diff I want to disable this by writing these 2 information directly in config
and hgsubversion will just read from there.

Reviewed By: mitrandir77

Differential Revision: D15989449

fbshipit-source-id: fd9cead094245b3c0baf6e7e1099cc302899668f