atomicetempfile: add `filesync` before closing
Summary:
Adding `filesync(fd)` (applies `flush()` and `os.fsync()` on fd) to make writes to the file more durable.

Before I ecountered the issue with on of the atomicfiles to have only zeros in it. markbt suggested that it could happend because of the system crashes, when the file size is still remembered but the content can be lost.
I first thought apply `filesync` only to that particular case, but decided to add into the `atomictempfile` implementation.

Reviewed By: mitrandir77

Differential Revision: D16078908

fbshipit-source-id: 5606b49aca766bf5f03fe4e763aa2c66c48ccac5