indexedlog: recover from interruptted log write
Summary:
In case an interruptted write (some data written to log, but meta is not
updated) happened, the previous behavior is to error out. However the callsite
does not actually have a reasonable way to recover from an incompleted write.

This diff changes the write behavior to simply ignore the physical length of
the "log" file but always use the length specified in "metadata". This should
make Log functional in those interruptted write cases.

Reviewed By: xavierd

Differential Revision: D16076658

fbshipit-source-id: c16b6d6fb942a438d4304bd927169b6dd919cb86